dashboard adjusted
