+++ app/javascript/plugins/init_tabs_click.js
@@ -1,0 +1,22 @@
+
+
+const tabs = document.querySelectorAll('.tab');
+
+const initTabsClick = () => {
+  if (tabs) {
+    tabs.forEach((tab) => {
+      tab.addEventListener('click', (event) => {
+        tabs.forEach((tab) => {
+          tab.classList.remove('active');
+          document.querySelector(`.${tab.id}`).style.display = 'none';
+        });
+        const activeTab = event.currentTarget;
+        activeTab.classList.add('active');
+        const activeDiv = document.querySelector(`.${activeTab.id}`);
+        activeDiv.style.display = 'block';
+      });
+    })
+  }
+}
+
+export { initTabsClick };
