+++ app/views/pages/dashboard.html.erb
@@ -5,6 +5,7 @@
     <li class="tab list-inline-item active" id="my-dogs">My dogs</li>
     <li class="tab list-inline-item" id="add-a-dog">Add a dog</li>
     <li class="tab list-inline-item" id="my-rental-requests">My rental requests</li>
+    <li class="tab list-inline-item" id="my-informations">My informations</li>
 
   </ul>
 
@@ -103,5 +104,12 @@
   <div class="my-rental-requests">
   </div>
 
+  <div class="my-informations">
+    <div class="row">
+      <div class="col-12 col-lg-6 offset-lg-3">
+
+      </div>
+    </div>
+  </div>
 </div>
